[REFACTOR] 마이 페이지 설정의 토글 파트 리팩토링
