+++ src/api/mypageApi.ts
@@ -1,5 +1,3 @@
-import type { queryData } from "@/components/mypage/article/MySettings";
-
 import { instance } from "./api";
 
 export const mypageApi = {
@@ -17,7 +15,7 @@ export const mypageApi = {
     const { data } = await instance.get(`/users/preferences`);
     return data;
   },
-  updateAlertSetting: async (payload: Partial<queryData>) => {
+  updateAlertSetting: async (payload: Partial<preferencesResponse>) => {
     return await instance.patch(`/users/preferences`, payload);
   },
 };
