+++ src/components/mypage/article/MySettings.tsx
@@ -1,59 +1,10 @@
 import React from "react";
-import { useMutation, useQuery } from "@tanstack/react-query";
 
 import Toggle from "@/components/designSystem/Toggle";
-import { QUERY_KEYS } from "@/constants/global/querykeys";
-import { mypageApi } from "@/api/mypageApi";
-
-export interface queryData {
-  darkmodeStatus: boolean;
-  neighborAlert: boolean;
-  commentAlert: boolean;
-  chatAlert: boolean;
-}
+import useSettingToggle from "@/hooks/mypageAPI/useSettingToggle";
 
 const MySettings = () => {
-  const { data } = useQuery<queryData>({
-    queryKey: [QUERY_KEYS.ALERTSETTING],
-    queryFn: () => mypageApi.getAlertSetting(),
-  });
-  const { mutate } = useMutation({
-    mutationFn: (payload: Partial<queryData>) =>
-      mypageApi.updateAlertSetting(payload),
-  });
-
-  const myOptions = [
-    {
-      text: "이웃 글 작성 알림",
-      isChecked: data?.neighborAlert,
-      onChange: (checked?: boolean) => {
-        mutate({ neighborAlert: !checked });
-      },
-    },
-    {
-      text: "댓글 알림",
-      isChecked: data?.commentAlert,
-      onChange: (checked?: boolean) => {
-        mutate({ commentAlert: !checked });
-      },
-    },
-    {
-      text: "토론방 알림",
-      isChecked: data?.chatAlert,
-      onChange: (checked?: boolean) => {
-        mutate({ chatAlert: !checked });
-      },
-    },
-    {
-      text: "이웃 비공개",
-      isChecked: data?.darkmodeStatus,
-      onChange: (checked?: boolean) => {
-        mutate({ darkmodeStatus: !checked });
-      },
-      subDescription:
-        "* 이웃 비공개 시, 내가 추가한 이웃과 나를 추가한 이웃이 다른 사람에게 보이지 않습니다.",
-    },
-  ];
+  const { myOptions } = useSettingToggle();
 
   return (
     <div className={`px-[54px] py-[40px] w-full h-full overflow-auto`}>
